Add files via upload
Compliments, complaints and other forms of feedback provide valuable information regarding client satisfaction and provide us with an opportunity to improve upon all aspects of our service. Obtaining feedback is taken seriously and is seen as an opportunity for improvement.
We seek to resolve complaints from clients and staff at the earliest opportunity in a way that respects and values the person’s feedback.   The CCMS approach aims to be responsive, effective and fair.